+++ CHANGELOG.md
@@ -1,8 +1,26 @@
 ## Release (2025-XX-YY)
 - `cdn`: [v0.2.0](services/cdn/CHANGELOG.md#v020-2025-04-01)
   - **API enhancement:** Provide waiter infrastructure
+- `logme`: [v0.21.2](services/logme/CHANGELOG.md#v0212-2025-04-02)
+  - **Bugfix:** `PartialUpdateInstanceWaitHandler` does not finish when update is succeeded
+  - **Deprecation:** Deprecated `InstanceStateSuccess`, `InstanceStateFailed`, `InstanceTypeCreate`, `InstanceTypeUpdate`, `InstanceTypeDelete` and will be removed after 2nd October 2025
+- `mariadb`: [v0.21.2](services/mariadb/CHANGELOG.md#v0212-2025-04-02)
+  - **Bugfix:** `PartialUpdateInstanceWaitHandler` does not finish when update is succeeded
+  - **Deprecation:** Deprecated `InstanceStateSuccess`, `InstanceStateFailed`, `InstanceTypeCreate`, `InstanceTypeUpdate`, `InstanceTypeDelete` and will be removed after 2nd October 2025
+- `opensearch`: [v0.20.2](services/opensearch/CHANGELOG.md#v0202-2025-04-02)
+  - **Bugfix:** `PartialUpdateInstanceWaitHandler` does not finish when update is succeeded
+  - **Deprecation:** Deprecated `InstanceStateSuccess`, `InstanceStateFailed`, `InstanceTypeCreate`, `InstanceTypeUpdate`, `InstanceTypeDelete` and will be removed after 2nd October 2025
+- `redis`: [v0.21.2](services/redis/CHANGELOG.md#v0212-2025-04-02)
+  - **Bugfix:** `PartialUpdateInstanceWaitHandler` does not finish when update is succeeded
+  - **Deprecation:** Deprecated `InstanceStateSuccess`, `InstanceStateFailed`, `InstanceTypeCreate`, `InstanceTypeUpdate`, `InstanceTypeDelete` and will be removed after 2nd October 2025
+- `rabbitmq`: [v0.21.2](services/rabbitmq/CHANGELOG.md#v0212-2025-04-02)
+  - **Bugfix:** `PartialUpdateInstanceWaitHandler` does not finish when update is succeeded
+  - **Deprecation:** Deprecated `InstanceStateSuccess`, `InstanceStateFailed`, `InstanceTypeCreate`, `InstanceTypeUpdate`, `InstanceTypeDelete` and will be removed after 2nd October 2025
+
+
 
 ## Release (2025-03-27)
+
 - `alb`: [v0.2.1](services/alb/CHANGELOG.md#v021-2025-03-27)
   - **Bugfix:** Removed ConfigureRegion() from API client
 - `cdn`: [v0.1.1](services/cdn/CHANGELOG.md#v011-2025-03-27)